Fix compile errors

+++ auto_encoders.h
@@ -10,8 +10,8 @@ private:
     double distance;
     Encoder* left_encoder;
     Encoder* right_encoder;
-    int get_avg_ticks();
-    double convert_distance(int);
+    double get_avg_ticks();
+    double convert_distance(double);
     static const double WHEEL_RADIUS = 8.0/2;
     static const double TICKS_PER_REV = 250;
     static const double DRIVE_REDUCTION = 38.0/60;
